% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elephants.R
\docType{data}
\name{elephants}
\alias{elephants}
\title{The Ian Parker East African Elephant Data Set}
\format{

}
\source{
{The Ian Parker East African Elephant Data Sheets: A handbook for the transcribed biological data set. 2014.} \url{https://ufdc.ufl.edu/IR00004209/00001}

 See also:

Primary Source
\url{https://ufdc.ufl.edu/AA00013409/00001}

Finding Guide
\url{https://findingaids.uflib.ufl.edu/repositories/2/resources/1440/collection_organization#tree::archival_object_272921}

Data collected by Ian Parker and Wildlife Services, Ltd., and donated to the University of Florida by Ian S.C. Parker in 2010.  The data were then transcribed by Reboussin, Daniel and Norton, Hannah in 2013.

"The Ian Parker East African Elephant Data Sheets".
}
\usage{
data(elephants)
}
\description{
East African elephant (Loxodonta africana) biological data from 3,175 field record sheets (in some instances, multiple cases were assigned to a single sheet). The data were recorded from 1965 to 1969 during herd culling designed to mitigate elephant herd overpopulation at multiple environmentally stressed national park sites. The data are unique in several ways: the conditions in which such a large number of elephants were represented are unlikely to be reproduced; the sampling represents natural (albeit stressed) family groups rather than trophies or weak animals (as is generally true in managed environments). From The Ian Parker East African Elephant Data Sheets: A handbook for the transcribed biological data set.
}
\keyword{datasets}
