% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ea_elephants.R
\docType{data}
\name{ea_elephants}
\alias{ea_elephants}
\title{The Ian Parker East African Elephant Data Set}
\format{
A data frame with 3307 observations on the following 100 variables.
\itemize{
\item{\code{Template_ID}}{ Internal variable indicates REDCap online template version employed by transcriber.}
\item{\code{lotSize}}{ size of lot (acres)}
\item{\code{age}}{ age of house (years)}
\item{\code{landValue}}{ value of land (US dollars)}
\item{\code{livingArea}}{ living are (square feet)}
\item{\code{pctCollege}}{ percent of neighborhood that graduated college}
\item{\code{bedrooms}}{ number of bedrooms}
\item{\code{firplaces}}{ number of fireplaces}
\item{\code{bathrooms}}{ number of bathrooms (half bathrooms have no shower or tub)}
\item{\code{rooms}}{ number of rooms}
\item{\code{heating}}{ type of heating system}
\item{\code{fuel}}{ fuel used for heating}
\item{\code{sewer}}{ type of sewer system}
\item{\code{waterfront}}{ whether property includes waterfront}
\item{\code{newConstruction}}{ whether the property is a new construction}
\item{\code{centralAir}}{ whether the house has central air}
}
}
\source{
Data collected by Candice Corvetti and used in the "Stat 101" case study 
"How much is a Fireplace Worth".  
See also 
<https://www.saratogacountyny.gov/departments/real-property-tax-service-agency/>
}
\usage{
data(ea_elephants)
}
\description{
East African elephant (Loxodonta africana) biological data from 3,175 field record sheets. The data were recorded from 1965 to 1969 during herd culling designed to mitigate elephant herd overpopulation at multiple environmentally stressed national park sites.
}
\keyword{datasets}
