% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elephants.R
\docType{data}
\name{elephants}
\alias{elephants}
\title{Body measurements for African and Asian elephants}
\format{
A tibble with 1700 rows and 9 variables
\describe{
\item{Category}{category (African, Asian)}
\item{Age_Category}{age category ("less than 10 years", "10-20 years", "greater than 20 years")}
\item{Gender}{elephant gender (male, female)}
\item{Flank_Girth}{flank girth in meters}
\item{Heart_Girth}{heart girth in meters}
\item{Neck_Girth}{neck girth in meters}
\item{Weight}{body weight in meters}
\item{Height}{shoulder height in meters}
\item{Fore_Feet_Circumference}{fore feet circumference in meters}
}
}
\source{
simulated dataset generated according to the findings of past studies
}
\usage{
elephants
}
\description{
Simulated dataset for body measurements of elephants
}
\keyword{datasets}
